+++ src/auth/auth.service.ts
@@ -13,7 +13,6 @@ import { CreateUserDto } from 'src/user/dto/create-user.dto';
 import { CreateOrganisationDto } from 'src/user/dto/create-organisation.dto';
 import { RegisterUserDto } from './dto/register.dto';
 import { CreateGoogleUserDto } from 'src/user/dto/create-google-user.dto';
-import { VerifyTokenDto } from 'src/user/dto/verify-token.dto';
 
 @Injectable()
 export class AuthService {
@@ -36,7 +35,7 @@ export class AuthService {
     } else {
       throw new BadRequestException('Invalid user type');
     }
-    const payload = { email: newUser.email, id: newUser.id };
+    const payload = { id: newUser.id };
 
     return {
       access_token: this.jwt.sign(payload),
@@ -45,7 +44,8 @@ export class AuthService {
 
   async login(loginDto: LoginDto) {
     const { email, password } = loginDto;
-    const user = await this.userService.findOneByEmail(email);
+    const formattedEmail = email.toLowerCase();
+    const user = await this.userService.findOneByEmail(formattedEmail);
     if (!user) {
       throw new NotFoundException('User not found');
     }
@@ -53,7 +53,7 @@ export class AuthService {
     if (!isPasswordValid) {
       throw new BadRequestException('Invalid credentials');
     }
-    const payload = { email: user.email, id: user.id };
+    const payload = { id: user.id };
     return {
       access_token: this.jwt.sign(payload),
     };
@@ -79,22 +79,19 @@ export class AuthService {
     if (!user) {
       throw new NotFoundException('User not found');
     }
-    const payload = { email: user.email, id: user.id };
+    const payload = { id: user.id };
     const token = this.jwt.sign(payload);
     return { access_token: token };
   }
 
-  async verifyToken(verifyTokenDto: VerifyTokenDto) {
-    const { id, email } = verifyTokenDto;
-    const user = await this.userService.findOne(id);
+  async verifyToken(id: string) {
+    const user = await this.prismaservice.user.findUnique({ where: { id } });
     if (!user) {
       throw new NotFoundException('User not found');
     }
     if (user.active !== true) {
       throw new UnauthorizedException('User not active');
     }
-    if (user.email === email && user.active === true) {
-      return true;
-    }
+    return { message: true };
   }
 }