+++ src/user/user.service.ts
@@ -11,15 +11,20 @@ import { PrismaService } from 'src/db/prisma.service';
 import * as argon2 from 'argon2';
 import { CreateOrganisationDto } from './dto/create-organisation.dto';
 import { CreateGoogleUserDto } from './dto/create-google-user.dto';
+import { UpdatePasswordDto } from './dto/update-password.dto';
 
 @Injectable()
 export class UserService {
   constructor(private readonly prismaService: PrismaService) {}
 
   async createUser(createUserDto: CreateUserDto) {
-    const { email, password } = createUserDto;
+    const { email, password, firstname, lastname } = createUserDto;
+    const formattedFirst =
+      firstname.charAt(0).toUpperCase() + firstname.slice(1);
+    const formattedLast = lastname.charAt(0).toUpperCase() + lastname.slice(1);
+    const formattedEmail = email.toLowerCase();
     const existingUser = await this.prismaService.user.findUnique({
-      where: { email },
+      where: { email: formattedEmail },
     });
 
     if (existingUser) {
@@ -29,16 +34,15 @@ export class UserService {
 
     const newUser = await this.prismaService.user.create({
       data: {
-        email: createUserDto.email,
+        email: formattedEmail,
         password: hashedPassword,
-        firstname: createUserDto.firstname,
-        lastname: createUserDto.lastname,
+        firstname: formattedFirst,
+        lastname: formattedLast,
         birthdate: new Date(createUserDto.birthdate),
         sex: createUserDto.sex,
         phone: createUserDto.phone,
         image_url: createUserDto.image_url,
         bio: createUserDto.bio,
-        city: createUserDto.city,
       },
     });
 
@@ -47,8 +51,9 @@ export class UserService {
 
   async createOrganisation(createOrganisationDto: CreateOrganisationDto) {
     const { email, password } = createOrganisationDto;
+    const formattedEmail = email.toLowerCase();
     const existingOrganisation = await this.prismaService.user.findUnique({
-      where: { email },
+      where: { email: formattedEmail },
     });
     if (existingOrganisation) {
       throw new ConflictException('User already exists');
@@ -56,15 +61,13 @@ export class UserService {
     const hashedPassword = await argon2.hash(password);
     const newOrganisation = await this.prismaService.user.create({
       data: {
-        // Ajout de 'data'
-        email: createOrganisationDto.email,
+        email: formattedEmail,
         password: hashedPassword,
         name: createOrganisationDto.name,
         phone: createOrganisationDto.phone,
         image_url: createOrganisationDto.image_url,
         bio: createOrganisationDto.bio,
-        city: createOrganisationDto.city,
-        type: 'ORGANISATION', // Vous voudrez probablement définir le type aussi
+        type: 'ORGANISATION',
       },
     });
     return newOrganisation;
@@ -105,7 +108,25 @@ export class UserService {
   }
 
   async findAll() {
-    const users = await this.prismaService.user.findMany();
+    const users = await this.prismaService.user.findMany({
+      select: {
+        id: true,
+        type: true,
+        email: true,
+        firstname: true,
+        lastname: true,
+        name: true,
+        birthdate: true,
+        sex: true,
+        phone: true,
+        image_url: true,
+        bio: true,
+        provider: true,
+        active: true,
+        updated_at: true,
+        created_at: true,
+      },
+    });
 
     if (!users) throw new NotFoundException('No users found');
 
@@ -115,6 +136,20 @@ export class UserService {
   async findOne(id: string) {
     const existingUser = await this.prismaService.user.findUnique({
       where: { id },
+      select: {
+        id: true,
+        type: true,
+        email: true,
+        firstname: true,
+        lastname: true,
+        name: true,
+        birthdate: true,
+        sex: true,
+        phone: true,
+        bio: true,
+        provider: true,
+        active: true,
+      },
     });
 
     if (!existingUser) throw new NotFoundException('User not found');
@@ -141,10 +176,13 @@ export class UserService {
   }
 
   async update(id: string, updateUserDto: UpdateUserDto) {
+    const { email } = updateUserDto;
     const existingUser = await this.findOne(id);
 
-    if (!existingUser) {
-      throw new NotFoundException('User not found');
+    const emailExists = await this.findOneByEmail(email);
+
+    if (emailExists) {
+      throw new ConflictException('Email is already taken');
     }
 
     const updatedUser = await this.prismaService.user.update({
@@ -155,7 +193,38 @@ export class UserService {
     });
 
     if (updatedUser) {
-      ('User updated successfully');
+      return { message: 'User updated successfully' };
+    } else {
+      throw new BadRequestException('User not updated');
+    }
+  }
+
+  async updatePassword(id: string, updatePasswordDto: UpdatePasswordDto) {
+    const { oldPassword, newPassword } = updatePasswordDto;
+    const existingUser = await this.prismaService.user.findUnique({
+      where: { id },
+    });
+
+    const isPasswordValid = await argon2.verify(
+      existingUser.password,
+      oldPassword,
+    );
+
+    if (!isPasswordValid) {
+      throw new BadRequestException('Invalid credentials');
+    }
+
+    const hashedPassword = await argon2.hash(newPassword);
+
+    const updatedUser = await this.prismaService.user.update({
+      where: { id },
+      data: {
+        password: hashedPassword,
+      },
+    });
+
+    if (updatedUser) {
+      return { message: 'User password updated successfully' };
     } else {
       throw new BadRequestException('User not updated');
     }